feat: fine tune animations
